Fix listener start and stop logic
I changed the start listener from using a `with` block to starting and
stopping it in different methods, but this caused the listener to
automatically close directly after it was started.

This was because when I used the `with` statement, it would start the
listener thread, then I would join the thread, and the thread would stop
once you exited the `with` statement. But once I removed the `with`
statement, I had to use start and join to begin listening, and then stop
to terminate the listener.