New notifier file
I created a notifier class that can trigger system notifications. I did
this because of I'm handling a certain app behaviour a different way.
Anytime you try to add a song that's a duplicate, it triggers a system
notification that it's a duplicate song. It doesn't add the song
(because it's a duplicate) but it still saves it to the last_save.
The reason for this is so that it gives you ability to remove it for
some reason, which honestly idk why that makes sense if you were just
trying to add it.