Major changes and clean up, check description
First I renamed the buttons.py file to actions.py across all files, and
the variables within it. This just made a wayy more sense because the
name much more accurately reflects the purpose of the file/variables.

A big change I made was take all the keyboard.KeyCodes() and put them
into class constants, like `SAVE_MAIN = keyboard.KeyCode(ch...`.
This was much cleaner because I used the keyboard module many times and
it was also super long statements, so it made way more sense to store
the button keycodes in class constants.

Last change was that I renamed the `start_keyboard_listener` and
`stop_keyboard_listener` to `start_listener` and `stop_listener.
This way when I write my RasPi class I can write the start and stop
methods with the same name, so that when I reference the listeners in the
main QuickSaver component I don't have to change the code to work.
Because the listeners are called in QuickSaver methods, so if I wrote it
using `listener.start_keyboard_listener`, then I would have to change it
if I was to instead pass the RasPi listener.