Add duplicate track return
Instead of just returning None for both `no currently playing track` and
`duplicate track`, I added a `IS_DUPE` constant and return that when a
duplicate track is attempted to be added. I do this so that when the
QuickSaver component gets the response, it can trigger a duplicate song
notification and act accordingly. I still save the details to
`last_save` to give the user a chance to remove the original add of the
song if they don't want it in the playlist at all.