Major changes and clean up, check description
Minor Changes:
- Remove listener imports as they're not used here and actually passed
  as a parameter when initializing QuickSaver in the main app.py file
- Moved `start_listener` to new method so that when QuickSaver is
  initialized, there's actually something to run. When it was called
inside `__init__`, there was nothing to use QuickSaver for besides
initializing it, because everything runs automaticall inside.

Major Design/UX Change:
I decided to change up a bit how duplicate songs were dealt with.
When a duplicate song is attempted to be added, obviously we don't add
it to the playlist, the local track list, or the track log; because we
never added it. BUT, we do still save those details to `last_save`.

Now it's important for the user to know that the song they tried to add
was not actually added, so I created a notifier class that can trigger
system Notifications. When a user tries to add a duplicate song,
the duplicate song notification is triggered. The notification message
also says if they would like to remove the song, they can press undo.
This is possible because we still saved the information to the
`last_save` attribute, so when they press undo, even though the song
wasn't just added, it is still removed from when it existed in the
playlist from before.